refactor(api)!: tighten visibility and promote sandbox to public API
- Make internal constants pub(crate): CHUNK_OVERLAP_TOKENS, gpu_pool_and_normalize
- Remove deprecated f32_as_bytes helper; downstream uses bytemuck::cast_slice directly
- Relocate recency_decay from storage to retrieval module
- Promote sandbox module from hidden to official public API for MLX-driving consumers
- Update README.md to reflect API changes

Closes #103
